+++ src/api/instance.ts
@@ -48,7 +48,7 @@ export interface InstanceInfo {
   contact_account: Profile | null,
   authentication_methods: AuthenticationMethod[],
   login_message: string;
-  allow_unauthenticated?: {
+  allow_unauthenticated: {
     timeline_local: boolean,
   },
   blockchains: BlockchainInfo[];